% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohmyR-package.R
\docType{package}
\name{ohmyR-package}
\alias{ohmyR}
\alias{ohmyR-package}
\title{ohmyR: Oh this is my R package}
\description{
I don't know why I created this package with this name,
    but just to let you know this R package is completely for the ANGSD homework.
}
\author{
\strong{Maintainer}: Chi-Lam Poon \email{clpoon807@gmail.com}

}
\keyword{internal}
